Separate SearchableAction ITs into own class
